requirement.txt
img2pdf
pytest-playwright
requests
pyside6
pathlib
bs4